src/test/java/com/leetcode/strings/easy/Maximum69Number_1323Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class Maximum69Number_1323Test {

  private final Maximum69Number_1323 solution = new Maximum69Number_1323();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {

      return Stream.of(arguments(9669, 9969), arguments(9996, 9999), arguments(9999, 9999));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void maximum69Number(int input, int expectedResult) {

    int actualResult = solution.maximum69Number(input);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
